fix: reorganize files and fix PR comments
