+++ megaparse/core/parser/__init__.py
@@ -1,0 +1,4 @@
+from megaparse.core.parser.megaparser import MegaParser
+
+
+__all__ = ["MegaParser"]
