+++ tests/pdf/test_pdf_processing.py
@@ -4,11 +4,11 @@ from uuid import uuid4
 import pytest
 
 from langchain_openai import ChatOpenAI
-from megaparse.parser.llama import LlamaParser
-from megaparse.parser.type import StrategyEnum
-from megaparse.parser.unstructured_parser import UnstructuredParser
-from megaparse.parser.megaparse_vision import MegaParseVision
-from megaparse.main import MegaParse
+from megaparse.core.parser.llama import LlamaParser
+from megaparse.core.parser.type import StrategyEnum
+from megaparse.core.parser.unstructured_parser import UnstructuredParser
+from megaparse.core.parser.megaparse_vision import MegaParseVision
+from megaparse.core.megaparse import MegaParse
 
 
 def test_get_default_processors_megaparse():
